move `pfio/v2/_profiler.py` to `pfio/_profiler.py`
